requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
groq==0.15.0
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==0.8.11
Pillow==10.0.0
pytz==2023.3
httpx==0.24.1
python-jose==3.3.0
pytest==7.4.0
pytest-asyncio==0.21.1